requirements.txt
gradio
openai
python-dotenv
redis
